+++ src/themes/open/components/legacy/TheFooter.vue
@@ -64,6 +64,12 @@ export default {
 					to: "/imprint",
 					text: this.$t("components.legacy.footer.imprint"),
 				},
+        {
+          href: "/datenschutz",
+          text: this.$t("components.legacy.footer.privacy_policy_HPI"),
+          target: "_blank",
+          rel: "noopener",
+        },
 				{
 					href: this.$store.getters["filePaths/getSpecificFiles"]
 						.termsOfUseSchool,