+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Changed
 
+- SC-8887 - data privacy in footer links changed
 - SC-8448 - resolves Merlin URL before passing to legacy client
 - SC-9077 - refactored store actions returning data
 - SC-9126 - changed the way the stores state is accessed; `mapState` and `store.state` were substituted for `mapGetters` and `store.getters` respectively; and changed the naming of the template and several other stores getters